+++ examples/custom_wizard.rs
@@ -1,49 +1,48 @@
 use dialoguer::Confirm;
 use service_install::{install_user, Tense};
 
-fn main() {
+fn main() -> Result<(), Box<dyn std::error::Error>> {
     let steps = install_user!()
-        .current_exe()
-        .unwrap()
+        .current_exe()?
         .name("cli")
         .on_boot()
-        .prepare_install()
-        .unwrap();
+        .prepare_install()?;
 
     let mut canceld = false;
     let mut rollback_steps = Vec::new();
     for mut step in steps {
         if !Confirm::new()
-            .with_prompt(format!("{}?", step.describe(Tense::Question)))
-            .interact()
-            .unwrap()
+            .with_prompt(format!("{}?", step.describe(Tense::Active)))
+            .interact()?
         {
             canceld = true;
             break;
         }
-        if let Some(rollback) = step.perform().unwrap() {
+        if let Some(rollback) = step.perform()? {
             rollback_steps.push(rollback);
         }
     }
 
     if !canceld {
-        return;
+        return Ok(());
     }
 
     if rollback_steps.is_empty() {
         println!("Install aborted, no changes have been made");
-        return;
+        return Ok(());
     } else {
         if Confirm::new()
             .with_prompt("Install aborted, do you want to roll back any changes made?")
-            .interact()
-            .unwrap()
+            .interact()?
         {
             for step in &mut rollback_steps {
-                let did = step.describe();
-                step.perform().unwrap();
+                let did = step.describe(Tense::Present);
+                step.perform()?;
                 println!("{}", did);
             }
         }
     }
+
+    // lets remove the install to prevent polluting the system
+    install_user!().name("cli").prepare_remove()?.best_effort_remove()?;
 }