+++ src/install/init/cron/teardown.rs
@@ -113,7 +113,7 @@ impl RemoveStep for RemoveInstalled {
             Tense::Past => "Removed",
             Tense::Present => "Removing",
             Tense::Future => "Will remove",
-            Tense::Question => "Remove",
+            Tense::Active => "Remove",
         };
         let user = self
             .user
@@ -128,7 +128,7 @@ impl RemoveStep for RemoveInstalled {
             Tense::Past => "Removed",
             Tense::Present => "Removing",
             Tense::Future => "Will remove",
-            Tense::Question => "Remove",
+            Tense::Active => "Remove",
         };
         let user = self
             .user